nightly_sanity/nightly_sanity.sh: option for specifying the branch
Now the script looks into /root/branch file for deciding which branch
it should checkout to and pull. If the branch mentioned in the file
is not created it will try to create that branch. If wrong branch is
given or if any error happens while checking out the branch, then
script by default continues on master.