helper_scripts/install_parallel_glusterfs: script to install the glusterfs simultaneously on multiple machines
This scripts looks into a file to know the IP address of the machines
where glusterfs has to be installed. Then executes the download_and_install.sh
script on all the machines. It is assumed that download_and_install.sh
script is available on all the machines. glusterfs version is given
as the argument.