sanity/nightly_sanity/nightly_updated.sh: make changes for the copying the new statedump files
Previously statedump files were located in /tmp/with the name glusterfs.<pid>
Now (glusterfs-3.3.0qa12 onwards) its not like that. Brick statedumps are saved
in the pattern "brick-path.<pid>.dump" and client statedumps are saved in the
pattern glusterdump.<pid>.dump. Changes are done with this commit to make nightly
sanity script aware of that.