ping_pong: spawn multiple threads, which will do locking on the same fd
This patch provides an option for the user to give the number of threds
as an argument. Before, ping_pong functionlity (i.e. locking the byte range)
was done in a separate single thread. With this patch depending upon the
input argument, multiple threads will be spawned all of which will try
to do the locking using the same fd.

Change-Id: Ib68bef407d75351664a76e81474814bd703d285c
Signed-off-by: Raghavendra Bhat <raghavendrabhat@gluster.com>