nightly_sanity: make changes to enable geo-replication and quota
geo-replication and quota were not enabled during  the tests, and if
they were needed they had to be hardcoded in the script. Instead let the
user who is running the script decide whether to enable any of them (or
both of them), by giving arguments at the time of executing the script.

Change-Id: I228eab1b9df4a4c261fb48c33618b9b56801a3d9
Signed-off-by: Raghavendra Bhat <raghavendrabhat@gluster.com>