arequal: take arguments from user to ignore directories for checksum calculation
With changes such as gfid based backend in glusterfs, we have to ignore some
hidden directories from the backend (which creates the hidden directories for
better management of the volume). So this patch takes care of taking the list
of ignored directories from user as an input argument. Those directories which
are present in the path given by the user (of which arequal is calculating the
checksum), are not considered while crawling the path.

Change-Id: If6b2209088300c5c59a8b0fb74e854f3ce030571
Signed-off-by: Raghavendra Bhat <raghavendrabhat@gluster.com>